requirements.txt
# torch>=1.12.0
transformers>=4.26.1
sentence-transformers>=2.2.2
dp-transformers
peft>=0.3.0
numpy>=1.23.5
pandas>=1.5.3
scikit-learn>=1.2.2
tqdm>=4.65.0
matplotlib>=3.7.1
nltk>=3.8.1
spacy>=3.5.1
pyyaml>=6.0
joblib>=1.2.0
pillow>=9.4.0
requests>=2.28.2
mauve-text>=0.3.0
tensorboard>=2.12.0
datasets>=2.10.1
accelerate>=0.18.0
retry
openai
huggingface_hub